+++ projects/angular2-draggable/src/lib/models/resize-event.ts
@@ -1,0 +1,11 @@
+import { ISize } from './size';
+
+export interface IResizeEvent {
+  host: any;
+  handle: any;
+  size: ISize;
+  position: {
+    top: number;
+    left: number;
+  };
+}