+++ src/app/views/resizable-demo/resizable-demo.module.ts
@@ -4,6 +4,7 @@ import { SharedModule } from '../shared/shared.module';
 
 import { ResizableDemoRoutingModule } from './resizable-demo-routing.module';
 import { ResizeDefaultDemoComponent } from './resize-default-demo/resize-default-demo.component';
+import { ResizeEventDemoComponent } from './resize-event-demo/resize-event-demo.component';
 
 @NgModule({
   imports: [
@@ -12,7 +13,8 @@ import { ResizeDefaultDemoComponent } from './resize-default-demo/resize-default
     SharedModule
   ],
   declarations: [
-    ResizeDefaultDemoComponent
+    ResizeDefaultDemoComponent,
+    ResizeEventDemoComponent
   ]
 })
 export class ResizableDemoModule { }