+++ src/assets/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0-beta.1 (2018-06-26)
+
+#### New
++ ngResizable: Provide `(rzStart)`, `(rzResizing)`, `(rzStop)` event emitters
++ ngResizable: Provide `resetSize()`, `getStatus()` methods
+
 ## 2.0.0-beta.0 (2018-06-25)
 
 #### New